:recycle: refactor(ppddm-agent, ppddm-core): Move BoostedModelManager in ppddm-agent to ppddm-core as Predictor
