+++ ppddm-agent/src/main/scala/ppddm/agent/controller/dm/algorithm/DataMiningAlgorithm.scala
@@ -5,11 +5,12 @@ import java.util.Base64
 
 import com.typesafe.scalalogging.Logger
 import org.apache.spark.ml.PipelineModel
-import org.apache.spark.sql.DataFrame
+import org.apache.spark.sql.{DataFrame, SparkSession}
 import org.zeroturnaround.zip.ZipUtil
-import ppddm.agent.controller.dm.{BoostedModelManager, StatisticsManager}
+import ppddm.agent.controller.dm.StatisticsManager
 import ppddm.agent.exception.DataMiningException
 import ppddm.agent.store.DataStoreManager
+import ppddm.core.ai.Predictor
 import ppddm.core.rest.model.{Agent, AgentAlgorithmStatistics, Algorithm, AlgorithmName, BoostedModel, WeakModel}
 
 import scala.concurrent.Future
@@ -17,6 +18,7 @@ import scala.concurrent.ExecutionContext.Implicits.global
 
 trait DataMiningAlgorithm {
 
+  implicit val sparkSession: SparkSession = ppddm.agent.Agent.dataMiningEngine.sparkSession
   protected val logger: Logger = Logger(this.getClass)
 
   protected val agent: Agent // on which aAgent this DataMiningAlgorithm is running now
@@ -64,7 +66,7 @@ trait DataMiningAlgorithm {
         (weakModel.weight.get, pipelineModel.transform(dataFrame))
       }
 
-      val testPredictionDF = BoostedModelManager.predictWithWeightedAverageOfPredictions(testPredictionTuples)
+      val testPredictionDF = Predictor.predictWithWeightedAverageOfPredictions(testPredictionTuples)
 
       // Calculate statistics
       val statistics = StatisticsManager.calculateBinaryClassificationStatistics(testPredictionDF)
